Turn based system (#63)
Closes #52 

Turn passes around between players and only the current one can end
their turn.

I've decided to drop VContainer and dependency injection in general for
this project. The fact that this is a multiplayer game, and objects are
added dynamically using FishNet makes it way harder to implement, so
I'll be using global variables for access and will be focusing on using
R3 and UniTask instead. I have it working correctly here, but it needs a
refactoring.

https://github.com/user-attachments/assets/2727aaf5-1d79-4b32-b6c7-b8bfbd91f727