Pass a turn by selecting a module (#65)
Closes #54 

The player with the current turn can click on a module to pass to the
next one. Every player is able to see what the other player's selection
is using an outline around the module. The modules despawn when they are
clicked on and the player can't click on another until it is their turn
again.

https://github.com/user-attachments/assets/e1b60c4a-3a21-4eaf-b117-03173aef5e64