Be able to rematch the game (#77)
Closes #62 

When a player reaches a certain amount of wins, the game stops just like
when a round ends. The players are notified that the match has ended and
tells them if they've won or lost. A button appears on the host's screen
to start a new match with the win counts being reset.

https://github.com/user-attachments/assets/eb3275d1-3c6b-4060-900f-0a20ef0a60da