Refactoring prototype (#82)
Closes #78 

## Introduction

This task was created after I finished building the original simple
protoype for the game. I was pretty scared that if I kept building on it
as it was, I would end up with a very big mess that is impossible to
extend and edit. The idea was to restructure the code completely so that
it followed SOLID principles and use good architectural patterns.

## The struggle

I've been using AI to help me fix problems in programming for quite a
while now, so I thought it would be good at analyzing and refactoring a
project... turns out it's a disaster. I spent most of the 11 hours
struggling with the things that the models gave to me. I tried multiple
things:

- Using GitHub Copilot agent to refactor it -> Project didn't compile
and had a ton of errors.
- Turning the codebase into a .txt and give it to ChatGPT to tell me
what to do -> Didn't understand the limitations of multiplayer games.
- Sending an image of Rider's type dependency diagram to analyze -> Gave
answers that didn't make much sense since it didn't understand what it
meant to do.
- Exporting the diagram into PlantUML and sending it to ChatGPT -> Ended
up just removing some classes and not making a lot of changes.

## Coming to a conclusion

After struggling a lot trying to use AI for refactoring, I decided that
it just isn't a good enough tool to do architectural changes with and
ended up defining the following:

> - **Dependency Injection:** Client-Only and Server-Only things that
aren't spawned with Fish-Net.
> - **EventBus:** Communication between multiple modules/components.
> - **Model-View-Controller:** UI elements/Interactables.
> - **Scriptable Objects:** Configurations and Values.
> - **State Machine:** AI behaviors/Game states (Probably not needed). 
> 
> We will group by modules and make sure there is no coupling between
things that aren't directly related. We'll analyze classes that get too
big and try to break them up into multiple smaller classes.

## Results

The following are type dependency diagrams from before and after the
refactoring:

Before:

![ArchitechureBefore](https://github.com/user-attachments/assets/6c27538e-2ce8-427d-a37f-e16711a52137)

After:
![Architechture
After](https://github.com/user-attachments/assets/94053dae-3c53-4dbf-aac6-61994c60b722)

## Going forward

While not perfect, the architecture definitely improved a lot this week
and I did leaned a lot of useful things, like implementing an event bus,
how dependency injection doesn't work very well int networked objects,
some of FishNet's limitations, etc. Most importantly, I understood that
the most important thing is to do things a bit slower to think on how to
do things and to trust my own abilities. Yes, it is true that I'm a
relatively novice game programmer, but my logical thinking is excellent
and, while still important, a good architecture is not as vital here as
it is in a AAA game-as-a-service product. Sure, having bad architecture
would still slow down the project, but it won't be too hard to maintain.
If at any point it grows too big and messy to handle, then I can ask for
help and take a week or two to refactor it again, but the most probable
thing is that I will end up noticing how some things aren't optimal and
improve them as I go using my newfound insight on how I want to
structure the project.

# Let's keep going!