Lobby with Steam and localhost joining (#86)
Closes #83 

Before joining the match, a player can start hosting a lobby that a
maximum of 3 other players can join. When all the players are ready, the
host can start the match and they're transported into the game scene.

There is one scene for the Steam lobby creation and another to serve as
the "waiting room" for testing in the same machine via localhost. We are
using Heathen's Toolkit for Steamworks 2025 to handle the lobby
connections and events via the included LobbyManager. The connected
players are shown as user cards, the UI will change in the future, but
the functionality is there.


![image](https://github.com/user-attachments/assets/7a00c865-cd93-4d94-9fc6-f034eb4de296)