Players get more clues when a new round starts (#87)
Closes #74 

## The Fix
When a new round starts each player gets exactly one set of clues, no
matter how many there are in the match.

## The Error
The round start event was raised twice on the server, once on the server
rpc and once on the player's observers rpc. Since we are doing client
hosting we need to make sure that events are only raised once, either by
skipping on host or by not calling on the server.